interactive list functionalities added for to do list
